Reorganize store code
